fix datepicker styling issues between android and ios
